requirements.txt
altgraph==0.17.2
anyio==3.4.0
asgiref==3.4.1
asyncio==3.4.3
bcrypt==3.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.0.0
caer==1.9.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
ci-info==0.2.0
click==8.0.3
cmake==3.22.0
colorama==0.4.4
comtypes==1.1.10
configobj==5.0.6
configparser==5.2.0
cryptography==36.0.2
CurrencyConverter==0.16.11
deepspeech==0.9.3
dlib @ file:///C:/Users/shankarr/Documents/dlib-19.22.99-cp39-cp39-win_amd64.whl
ecdsa==0.17.0
etelemetry==0.3.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.70.0
filelock==3.7.0
fitz==0.0.1.dev2
Flask==2.1.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.4
httptools==0.2.0
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.1
jose==1.0.0
looseversion==1.0.1
lxml==4.8.0
MarkupSafe==2.1.1
MouseInfo==0.1.3
mypy==0.910
mypy-extensions==0.4.3
Naked==0.1.31
networkx==2.8.2
nibabel==3.2.2
nipype==1.8.1
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
passlib==1.7.4
pathlib==1.0.1
pefile==2021.9.3
Pillow==8.4.0
prov==2.0.0
py2exe==0.11.1.0
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.8.2
pydot==1.4.2
pydub==0.25.1
PyGetWindow==0.0.9
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyJWT==1.7.1
PyMsgBox==1.0.9
PyMuPDF==1.19.6
pyngrok==5.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pyttsx3==2.0
pytweening==1.0.4
pytz==2022.1
pywin32==302
pywin32-ctypes==0.2.0
pyxnat==1.4
PyYAML==6.0
rdflib==6.1.1
requests==2.26.0
rsa==4.8
scipy==1.8.1
shellescape==3.8.1
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
starlette==0.16.0
toml==0.10.2
traits==6.3.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
waitress==2.1.1
watchgod==0.7
websockets==10.1
Werkzeug==2.1.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
